e2e configuration
