+++ src/modules/items/items.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ItemsService } from './items.service';
 import { ItemsController } from './items.controller';
-import { PrismaService } from 'src/database/prisma/prisma.service';
+import { PrismaService } from '../../database/prisma/prisma.service';
 import { UploadFileService } from '../upload-file/upload-file.service';
 
 @Module({
